[MI-2372] Fixed issue #120 Modals opening on top of each other Changed the structure of webapp by removing all the reducers for modals and creating one reducer for global modal Modified all the required containers and components to use the new global modal state Made some new types and refactored all the types to move to new files
