+++ webapp/src/reducers/connectedState/index.test.ts
@@ -1,4 +1,4 @@
-import reducer, {setConnected, ConnectedState} from 'src/reducers/connectedState';
+import reducer, {setConnected} from 'src/reducers/connectedState';
 
 const previousState: ConnectedState = {
     connected: false,
