+++ webapp/src/reducers/refetchState/index.test.ts
@@ -1,4 +1,4 @@
-import reducer, {refetch, resetRefetch, RefetchState} from 'src/reducers/refetchState';
+import reducer, {refetch, resetRefetch} from 'src/reducers/refetchState';
 
 const previousState: RefetchState = {
     refetch: false,
