+++ webapp/src/types/common/store.d.ts
@@ -1,0 +1,42 @@
+type PluginState = RootState<{ [x: string]: QueryDefinition<void, BaseQueryFn<string | FetchArgs, unknown, FetchBaseQueryError, {}, FetchBaseQueryMeta>, never, void, 'pluginApi'>; }, never, 'pluginApi'>
+
+type ReduxState = {
+    'plugins-mattermost-plugin-servicenow': PluginState
+}
+
+type GlobalModalState = {
+    modalId: ModalId;
+    data?: EditSubscriptionData | CommentAndStateModalData | null;
+}
+
+type CommentModalState = {
+    open: boolean;
+    data?: CommentAndStateModalData;
+}
+
+type ConnectedState = {
+    connected: boolean;
+};
+
+type SubscriptionModalState = {
+    open: boolean;
+    data?: EditSubscriptionData;
+}
+
+type RefetchState = {
+    refetch: boolean;
+};
+
+type ShareRecordModalState = {
+    open: boolean;
+}
+
+type UpdateStateModalReduxState = {
+    open: boolean;
+    data?: CommentAndStateModalData;
+}
+
+type CommentAndStateModalData = {
+    recordType: RecordType;
+    recordId: string;
+}
